Update StartMainMenuOnLinux.sh
Small error was removed in the linux startup script. If you do want to clone, clone this repo to 'Arcade-4-Player-Games-kopie'. It is also possible to make an exact copy to 'Arcade-4-Player-Games-kopie' if you cloned this repo to another directory. It should run both on Windows and Linux in 'Arcade-4-Player-Games-kopie'.

Sorry for any inconvenience...

Cheers,

William Senn.