More standardized key(board)control.
Adjusted all files and programs for the latest version of processing, except for the MacOSX-version and the readme-files. Old readme-files can still be applied, except for the mentioned line-numbers. You will have to adjust those line-numbers a little.
I have tried to make the keyboard-controls more standardized on all games and menus, except for Pacman.
This also means that Linux versions do not need sudo anymore and these should be able to run under wsl (2), I did not test this yet...